==> xgboost/imputed/7_xgb_imputed_fairness.ipynb <==
# Fairness Assessment

import numpy as np
import pandas as pd
import pickle
 
import datetime
dte = datetime.datetime.now()
dte = dte.strftime("%Y%m%d")

# import custom functions
import sys
sys.path.append('/ML-AI-PCOR-2021/onc_functions')

from fairness import get_fairness_assessment

print('python-' + sys.version)

# load pooled results from the xgboost imputed model (not calibrated)

pred_df = pd.read_csv('./roc_results/2021_xgb_pooling_results_final_roc.csv')
pred_df = pred_df.loc[:,['averaged','died_in_90','usrds_id']].copy()
pred_df.head()

#load fairness details
with open('../complete_fairness_data.pickle', 'rb') as f:
        dataset = pickle.load(f)
dataset.head()

#combine these 2 dataframes
data = pred_df.merge(dataset, how='left', on=['usrds_id','died_in_90'])
data.head()

fairness = get_fairness_assessment(
                                    data,
                                    y_proba_col_name='averaged',
                                    y_true_col_name='died_in_90')
fairness.head(39)

#save (for the aucroc evaluated model)
fairness.to_csv('./roc_results/' + str(dte) + '_xgb_imputed_fairness.csv')
